Add: Reservation Controller

+++ tennis-reservation-system/src/main/java/skylarmaeve/tennisreservationsystem/dto/PriceResponseDto.java
@@ -2,10 +2,12 @@ package skylarmaeve.tennisreservationsystem.dto;
 
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public class PriceResponseDto {
-    private Integer price;
-    public PriceResponseDto(Integer price) {
+    private BigDecimal price;
+    public PriceResponseDto(BigDecimal price) {
         this.price = price;
     }
 }
